requirements.txt
# requirements.txt
# YOLOv8 번호판 인식 시스템 의존성

# 핵심 라이브러리
ultralytics>=8.0.0
opencv-python>=4.5.0
pytesseract>=0.3.8
Pillow>=8.0.0
numpy>=1.21.0

# 딥러닝 프레임워크
torch>=1.9.0
torchvision>=0.10.0

# 웹 프레임워크
flask>=2.0.0
gunicorn>=20.1.0

# 데이터 처리
pandas>=1.3.0
matplotlib>=3.3.0
seaborn>=0.11.0

# 설정 관리
pyyaml>=6.0
python-dotenv>=0.19.0

# 모니터링 및 메트릭
prometheus-client>=0.12.0
psutil>=5.8.0
GPUtil>=1.4.0

# 네트워킹 및 API
requests>=2.25.0
python-multipart>=0.0.5

# 백그라운드 작업
schedule>=1.1.0

# 개발 및 테스트
pytest>=6.0.0
pytest-cov>=3.0.0
black>=21.0.0
flake8>=4.0.0

# 고정확도 OCR 엔진들 (선택적 설치)
# 한국어 특화 OCR
# pororo>=1.0.0

# 다국어 고정확도 OCR
# paddleocr>=2.6.0
# paddlepaddle>=2.4.0

# 사용하기 쉬운 OCR
# easyocr>=1.6.0

# 선택적 의존성 (프로덕션 환경)
# PostgreSQL 사용시 주석 해제
# psycopg2-binary>=2.9.0

# Redis 사용시 주석 해제
# redis>=4.0.0